+++ src/main/java/dto/ResultRequest.java
@@ -1,4 +1,4 @@
-package domain;
+package dto;
 
 public class ResultRequest {
     private static final String LESS_THAN_MIN_LETTER_NUMBER_MESSAGE = "결괏값은 최소 1자이상 입력해야 합니다. ";
