+++ src/main/java/dto/ResultsRequest.java
@@ -1,4 +1,4 @@
-package domain;
+package dto;
 
 import java.util.ArrayList;
 import java.util.List;
